# reg_webapp/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ValidationIssueModel": {
        "description": "One validation issue \u2014 a 1:1 Pydantic wrapper of reg_schema's frozen\n``ValidationIssue`` dataclass. ``level`` is the tri-state severity; ``path`` is\nan RFC-6901 JSON pointer into ``project_data.json`` (empty for whole-document\nissues); ``code`` is the stable, namespaced rule identifier the SPA maps to a\nUI affordance.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "level": {
            "enum": [
              "error",
              "warning",
              "info"
            ],
            "title": "Level",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "level",
          "code",
          "path",
          "message"
        ],
        "title": "ValidationIssueModel",
        "type": "object"
      },
      "ValidationResultModel": {
        "description": "`POST /api/project/validate` response \u2014 the concatenated issue list\n(structural \u29fa block \u29fa semantic) plus the derived ``ok`` flag.\n\n``ok`` mirrors ``reg_schema.ValidationResult.ok``: True iff NO error-level\nissue is present (warnings/info do not flip it). A validation FAILURE is a\nSUCCESSFUL validation RESPONSE \u2014 this carries HTTP 200 with ``ok=false`` and\nthe issues; 4xx is reserved for a malformed request (bad JSON / oversized body\n/ wrong content-type), never for a spec that simply failed to validate.",
        "properties": {
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssueModel"
            },
            "title": "Issues",
            "type": "array"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "issues"
        ],
        "title": "ValidationResultModel",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "reg_webapp",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/project/order": {
      "post": {
        "description": "Render the steward's default v1 order-export CSV.\n\nReads the raw dict and runs the STRUCTURAL gate (see reg_schema/DESIGN.md \u2192\nStructural rules and issue codes) before rendering: the\n``ProjectData`` model enforces only field types, while the structural rules\n(FQID shape, period grammar, the binding/source-prefix match) live in\n``validate_structural`` \u2014 so a Pydantic-valid-but-structurally-invalid spec\n(e.g. a malformed ``register_variant`` or bad period token) would otherwise\nrender a bad provider order at 200. A structurally invalid spec \u2192 422.\n``async`` + ``run_in_threadpool`` (blocking display_name resolution off the\nevent loop), mirroring ``/validate``.",
        "operationId": "order_project_api_project_order_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/csv": {}
            },
            "description": "Successful Response"
          }
        },
        "summary": "Order Project"
      }
    },
    "/api/project/validate": {
      "post": {
        "description": "Validate a ``project_data.json``. Returns 200 with the concatenated\nstructural \u29fa block \u29fa semantic issue list + the derived ``ok`` flag; a 4xx is\nreserved for a malformed REQUEST (``read_raw_json_object`` / the body cap).\n\nThis is the SEMANTIC validator (reg_meta-backed). It now ALSO runs the\nbuild-time cross-block referential checks (orphan ``binding_options`` keys /\nsuppress_k-on-non-categorical) \u2014 that half of the old ``/validate``\u2194``/bundle``\ndivergence is CLOSED. The ONLY residual gap: ``/bundle`` additionally runs the\nstep-4 capability gates (e.g. a build-required ``display_name``), which\n``/validate`` does NOT \u2014 so a spec ``/validate`` greenlights can still 422 at\n``/bundle`` on a capability gate (an intentional lenient residual: ``/validate``\ndefaults ``display_name`` from reg_meta).\n\n``async`` only to read the body off the wire; the BLOCKING work (the structural\nparse + the semantic layer's per-binding sqlite resolution) is offloaded to the\nthreadpool via ``run_in_threadpool`` so it never stalls the event loop (the\ncatalog routes are plain ``def`` for the same reason). The reg_meta connection\nopens on that threadpool thread (one thread \u2192 the cross-thread sqlite P1 can't\nrecur).",
        "operationId": "validate_project_api_project_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResultModel"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Validate Project"
      }
    }
  }
}